+++ pages/api/xAPI/agent/profile.ts
@@ -5,7 +5,7 @@ import {
   agent,
   inverseFunctionalIdentifier,
 } from 'lib/types/x-api/actor'
-import { document } from 'lib/types/x-api/document'
+import { createEtag, document, mergeDocuments } from 'lib/types/x-api/document'
 import { z } from 'zod'
 import { prisma } from 'lib/server/db'
 
@@ -38,39 +38,55 @@ export default defineEndpoints({
         query: { agent, profileId, since },
       },
     }) => {
-      if (profileId) {
-        const prismaDocument = await prisma.document.findFirst({
-          where: {
-            profileId,
-            agent: inverseFunctionalIdentifier.parse(agent),
-          },
-        })
-        if (prismaDocument) {
-          let parsedDocument = document.parse(prismaDocument)
-          res.setHeader('content-type', 'application/json')
-          res.status(200).json(parsedDocument)
+      try {
+        res.setHeader('content-type', 'application/json')
+        if (profileId) {
+          const prismaDocument = await prisma.document.findFirst({
+            where: {
+              profileId,
+              agent: inverseFunctionalIdentifier.parse(agent),
+            },
+          })
+          if (prismaDocument) {
+            const newEtag = createEtag(prismaDocument)
+            res.setHeader('ETag', newEtag)
+            let parsedDocument = document.parse(prismaDocument)
+            res.status(200).json(parsedDocument)
+          } else {
+            res.status(404).end()
+          }
         } else {
-          res.status(404).end()
+          const prismaDocuments = await prisma.document.findMany({
+            where: {
+              agent: inverseFunctionalIdentifier.parse(agent),
+              timestamp: { gte: since },
+            },
+            select: {
+              profileId: true,
+              timestamp: true,
+            },
+          })
+          if (prismaDocuments) {
+            let parsedDocuments = prismaDocuments
+              .map((prismaDocument) => prismaDocument.profileId)
+              .filter((id): id is string => {
+                return !!id
+              })
+
+            let lastModified = prismaDocuments
+              .map((doc) => doc.timestamp)
+              .filter((id): id is Date => {
+                return !!id
+              })
+              .sort()
+              .reverse()[0]
+            res.setHeader('Last-Modified', lastModified.toISOString())
+            res.status(200).json(parsedDocuments)
+          }
         }
-      } else {
-        const prismaDocuments = await prisma.document.findMany({
-          where: {
-            agent: inverseFunctionalIdentifier.parse(agent),
-            timestamp: { gte: since },
-          },
-          select: {
-            profileId: true,
-          },
-        })
-        if (prismaDocuments) {
-          let parsedDocuments = prismaDocuments
-            .map((prismaDocument) => prismaDocument.profileId)
-            .filter((id): id is string => {
-              return !!id
-            })
-          res.setHeader('content-type', 'application/json')
-          res.status(200).json(parsedDocuments)
-        }
+      } catch (error) {
+        console.error(error)
+        res.status(400).end()
       }
     },
   },
@@ -93,6 +109,7 @@ export default defineEndpoints({
     handler: async ({
       res,
       req: {
+        headers: { 'If-None-Match': ifNoneMatch, 'If-Match': ifMatch },
         body,
         query: { agent, profileId },
       },
@@ -100,6 +117,17 @@ export default defineEndpoints({
       res.setHeader('content-type', 'application/json')
       try {
         const old = await prisma.document.findUnique({ where: { profileId } })
+        if (old !== null) {
+          const etag = createEtag(old)
+          if (ifNoneMatch === etag) {
+            res.status(304).end()
+            return
+          }
+          if (ifMatch && ifMatch !== etag) {
+            res.status(412).end()
+            return
+          }
+        }
         await prisma.document
           .create({
             data: {
@@ -150,6 +178,7 @@ export default defineEndpoints({
     handler: async ({
       res,
       req: {
+        headers: { 'If-None-Match': ifNoneMatch, 'If-Match': ifMatch },
         body,
         query: { agent, profileId },
       },
@@ -157,6 +186,17 @@ export default defineEndpoints({
       res.setHeader('content-type', 'application/json')
       try {
         const old = await prisma.document.findUnique({ where: { profileId } })
+        if (old !== null) {
+          const etag = createEtag(old)
+          if (ifNoneMatch === etag) {
+            res.status(304).end()
+            return
+          }
+          if (ifMatch && ifMatch !== etag) {
+            res.status(412).end()
+            return
+          }
+        }
         await prisma.document
           .create({
             data: {
@@ -204,39 +244,81 @@ export default defineEndpoints({
         schema: undefined,
       },
     ],
+
     handler: async ({
       res,
       req: {
+        headers: { 'If-Match': ifMatch },
+
         query: { agent, profileId },
       },
     }) => {
       res.setHeader('content-type', 'application/json')
-      if (profileId) {
-        try {
-          await prisma.document.deleteMany({
-            where: {
-              profileId,
-
-              agent: inverseFunctionalIdentifier.parse(agent),
-            },
-          })
-        } catch (error) {
-          console.error(error)
-          res.status(400).end()
-        }
-      } else {
-        {
-          try {
-            await prisma.document.deleteMany({
+      try {
+        if (profileId) {
+          if (ifMatch && ifMatch !== '*' && typeof ifMatch === 'string') {
+            const old = await prisma.document.findFirst({
               where: {
+                profileId,
                 agent: inverseFunctionalIdentifier.parse(agent),
               },
             })
-          } catch (error) {
-            console.error(error)
-            res.status(400).end()
+            if (old === null) res.status(400).end()
+            if (old !== null) {
+              const etag = createEtag(old)
+              if (ifMatch !== etag) {
+                res.status(412).end()
+                return
+              }
+            }
+            await prisma.document.deleteMany({
+              where: {
+                profileId,
+                agent: inverseFunctionalIdentifier.parse(agent),
+              },
+            })
+          } else {
+            if (ifMatch && ifMatch !== '*' && typeof ifMatch !== 'string') {
+              ifMatch.forEach(async (etag) => {
+                const old = await prisma.document.findFirst({
+                  where: {
+                    profileId,
+                    agent: inverseFunctionalIdentifier.parse(agent),
+                  },
+                })
+                if (old === null) res.status(400).end()
+                if (old !== null) {
+                  const newEtag = createEtag(old)
+                  if (etag !== newEtag) {
+                    res.status(412).end()
+                    return
+                  }
+                }
+                await prisma.document.deleteMany({
+                  where: {
+                    profileId,
+                    agent: inverseFunctionalIdentifier.parse(agent),
+                  },
+                })
+              })
+            }
+            await prisma.document.deleteMany({
+              where: {
+                profileId,
+                agent: inverseFunctionalIdentifier.parse(agent),
+              },
+            })
           }
+        } else {
+          await prisma.document.deleteMany({
+            where: {
+              agent: inverseFunctionalIdentifier.parse(agent),
+            },
+          })
         }
+      } catch (error) {
+        console.error(error)
+        res.status(400).end()
       }
     },
   },
